Update lang-zh-hans.json
@CoolLKKPS better?
Also, I talked with the guy who makes the interface I use and it is an incompatibility between it and the game translation mod. Chinese characters are larger and need more pixels to differentiate than other characters. I succeeded to make them fit, but they started to become blurry so it wasnt worth it